+++ src/TestCase.php
@@ -12,6 +12,7 @@ use ReflectionClass;
  * One test suit
  *
  * @author Jakub Konečný
+ * @property-read Job[] $jobs @internal
  */
 abstract class TestCase {
   use \Nette\SmartObject;
@@ -48,30 +49,32 @@ abstract class TestCase {
    * @return Job[]
    */
   protected function getJobs(): array {
-    $jobs = [];
-    $r = new ReflectionClass(static::class);
-    $methods = array_values(preg_grep(static::METHOD_PATTERN, array_map(function(\ReflectionMethod $rm) {
-      return $rm->getName();
-    }, $r->getMethods())));
-    foreach($methods as $method) {
-      /** @var callable $callback */
-      $callback = [$this, $method];
-      $job = [
-        "name" => $this->getJobName(static::class, $method),
-        "callback" => $callback,
-        "params" => [],
-        "skip" => $this->skipChecker->shouldSkip(static::class, $method),
-        "shouldFail" => $this->shouldFailChecker->shouldFail(static::class, $method),
-      ];
-      $data = $this->dataProvider->getData(static::class, $method);
-      if(count($data) > 0) {
-        foreach($data as $value) {
-          $job["params"][0] = $value;
+    static $jobs = [];
+    if(count($jobs) === 0) {
+      $r = new ReflectionClass(static::class);
+      $methods = array_values(preg_grep(static::METHOD_PATTERN, array_map(function(\ReflectionMethod $rm) {
+        return $rm->getName();
+      }, $r->getMethods())));
+      foreach($methods as $method) {
+        /** @var callable $callback */
+        $callback = [$this, $method];
+        $job = [
+          "name" => $this->getJobName(static::class, $method),
+          "callback" => $callback,
+          "params" => [],
+          "skip" => $this->skipChecker->shouldSkip(static::class, $method),
+          "shouldFail" => $this->shouldFailChecker->shouldFail(static::class, $method),
+        ];
+        $data = $this->dataProvider->getData(static::class, $method);
+        if(count($data) > 0) {
+          foreach($data as $value) {
+            $job["params"][0] = $value;
+            $jobs[] = new Job($job["name"], $job["callback"], $job["params"], $job["skip"], $job["shouldFail"]);
+            $job["params"] = [];
+          }
+        } else {
           $jobs[] = new Job($job["name"], $job["callback"], $job["params"], $job["skip"], $job["shouldFail"]);
-          $job["params"] = [];
         }
-      } else {
-        $jobs[] = new Job($job["name"], $job["callback"], $job["params"], $job["skip"], $job["shouldFail"]);
       }
     }
     return $jobs;