+++ src/Tester.php
@@ -3,8 +3,10 @@ declare(strict_types=1);
 
 namespace MyTester;
 
+use Ayesh\PHP_Timer\Timer;
 use Jean85\PrettyVersions;
 use MyTester\Bridges\NetteRobotLoader\TestSuitsFinder;
+use Nette\Utils\Finder;
 
 /**
  * Automated tests runner
@@ -17,6 +19,7 @@ final class Tester {
   use \Nette\SmartObject;
 
   private const PACKAGE_NAME = "konecnyjakub/mytester";
+  private const TIMER_NAME = "My Tester";
   
   /** @var string[] */
   private array $suits;
@@ -26,6 +29,9 @@ final class Tester {
   ];
   public ITestSuitFactory $testSuitFactory;
   private string $folder;
+  /** @var SkippedTest[] */
+  private array $skipped = [];
+  private string $results = "";
   
   public function __construct(string $folder) {
     $this->onExecute[] = [$this, "printInfo"];
@@ -56,8 +62,9 @@ final class Tester {
       if(!$suit->run()) {
         $failed = true;
       }
+      $this->saveResults($suit);
     }
-    Environment::printResults();
+    $this->printResults();
     exit((int) $failed);
   }
 
@@ -69,6 +76,80 @@ final class Tester {
     echo "\n";
     echo "PHP " . PHP_VERSION . "(" . PHP_SAPI . ")\n";
     echo "\n";
+  }
+
+  private function printResults(): void {
+    $results = $this->results;
+    echo $results . "\n";
+    $this->printSkipped();
+    $failed = str_contains($results, TestCase::RESULT_FAILED);
+    if(!$failed) {
+      echo "\n";
+      echo "OK";
+    } else {
+      static::printFailed();
+      echo "\n";
+      echo "Failed";
+    }
+    $resultsLine = " (" . strlen($results) . " tests";
+    if($failed) {
+      $resultsLine .= ", " . substr_count($results, TestCase::RESULT_FAILED) . " failed";
+    }
+    if(str_contains($results, TestCase::RESULT_SKIPPED)) {
+      $resultsLine .= ", " . substr_count($results, TestCase::RESULT_SKIPPED) . " skipped";
+    }
+    Timer::stop(static::TIMER_NAME);
+    $time = Timer::read(static::TIMER_NAME, Timer::FORMAT_HUMAN);
+    $resultsLine .= ", $time)";
+    echo $resultsLine . "\n";
+  }
+
+  /**
+   * Print info about skipped tests
+   */
+  private function printSkipped(): void {
+    foreach($this->skipped as $skipped) {
+      $reason = "";
+      if($skipped->reason) {
+        $reason = ": {$skipped->reason}";
+      }
+      echo "Skipped $skipped->name$reason\n";
+    }
+  }
+
+  /**
+   * Print info about failed tests
+   */
+  private function printFailed(): void {
+    $filenameSuffix = ".errors";
+    $files = Finder::findFiles("*$filenameSuffix")->in($this->folder);
+    /** @var \SplFileInfo $file */
+    foreach($files as $name => $file) {
+      echo "--- " . $file->getBasename($filenameSuffix) . "\n";
+      echo file_get_contents($name);
+    }
+  }
+
+  private function saveResults(TestCase $testCase): void {
+    $jobs = $testCase->jobs;
+    foreach($jobs as $job) {
+      switch($job->result) {
+        case Job::RESULT_PASSED:
+          $result = TestCase::RESULT_PASSED;
+          break;
+        case Job::RESULT_SKIPPED:
+          $result = TestCase::RESULT_SKIPPED;
+          $this->skipped[] = new SkippedTest($job->name, (is_string($job->skip) ? $job->skip : ""));
+          break;
+        case Job::RESULT_FAILED:
+          $result = TestCase::RESULT_FAILED;
+          break;
+        default:
+          $result = "";
+          break;
+      }
+      $this->results .= $result;
+    }
   }
 }
 ?>